Upload dedicated target benchmark results to Conbench
