+++ scripts/benchmark-requirements.txt
@@ -12,8 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# To be installed from the root of the repo
-
-conbench
-git+https://github.com/conbench/benchalerts.git@0.4.0
-./scripts/veloxbench
+benchadapt@git+https://github.com/conbench/conbench.git@a791fe6#subdirectory=benchadapt/python
+benchalerts@git+https://github.com/conbench/benchalerts.git@0.4.0
