--- scripts/veloxbench/.flake8
@@ -1,2 +1,0 @@
-[flake8]
-ignore = E501,E731,W503
